requirements.txt
asttokens==2.4.1
backports.functools-lru-cache==1.6.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cuquantum-cu12==23.6.1
custatevec-cu12==1.4.1
cutensor-cu12==1.7.0
cutensornet-cu12==2.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter-client==7.3.4
jupyter_core==5.5.0
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
mqt.ddsim==1.20.1
mycolorpy==1.5.1
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.3.2.9
nvidia-cuda-runtime-cu12==12.3.52
nvidia-cusolver-cu12==11.5.3.52
nvidia-cusparse-cu12==12.1.3.153
nvidia-nvjitlink-cu12==12.3.52
objprint==0.2.3
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3
platformdirs==4.0.0
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pylatexenc==2.10
pyparsing==3.1.1
pyspnego==0.10.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.0
qiskit==0.44.2
qiskit-aer==0.13.0
qiskit-aer-gpu==0.12.2
qiskit-algorithms==0.2.1
qiskit-ibm-provider==0.7.1
qiskit-terra==0.25.2.1
qtcodes==0.2.0
quantum-gates==1.0.4
requests==2.31.0
requests-ntlm==1.2.0
retworkx==0.13.2
rustworkx==0.13.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.0.0
six==1.16.0
stack-data==0.6.2
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
tornado==6.1
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
websocket-client==1.6.4
websockets==11.0.3
wheel==0.41.2
zipp==3.17.0
